+++ org/postgresql/jdbc2/Jdbc2Blob.java
@@ -3,7 +3,7 @@
 * Copyright (c) 2004-2005, PostgreSQL Global Development Group
 *
 * IDENTIFICATION
-*   $PostgreSQL: pgjdbc/org/postgresql/jdbc2/Jdbc2Blob.java,v 1.3 2004/11/09 08:49:03 jurka Exp $
+*   $PostgreSQL: pgjdbc/org/postgresql/jdbc2/Jdbc2Blob.java,v 1.4 2005/01/11 08:25:46 jurka Exp $
 *
 *-------------------------------------------------------------------------
 */
@@ -13,7 +13,7 @@ package org.postgresql.jdbc2;
 public class Jdbc2Blob extends AbstractJdbc2Blob implements java.sql.Blob
 {
 
-    public Jdbc2Blob(org.postgresql.PGConnection conn, int oid) throws java.sql.SQLException
+    public Jdbc2Blob(org.postgresql.PGConnection conn, long oid) throws java.sql.SQLException
     {
         super(conn, oid);
     }
