+++ org/postgresql/jdbc4/Jdbc4Clob.java
@@ -3,7 +3,7 @@
 * Copyright (c) 2004-2005, PostgreSQL Global Development Group
 *
 * IDENTIFICATION
-*   $PostgreSQL$
+*   $PostgreSQL: pgjdbc/org/postgresql/jdbc4/Jdbc4Clob.java,v 1.1 2006/06/08 10:34:52 jurka Exp $
 *
 *-------------------------------------------------------------------------
 */
@@ -13,7 +13,7 @@ package org.postgresql.jdbc4;
 public class Jdbc4Clob extends AbstractJdbc4Clob implements java.sql.Clob
 {
 
-    public Jdbc4Clob(org.postgresql.PGConnection conn, int oid) throws java.sql.SQLException
+    public Jdbc4Clob(org.postgresql.PGConnection conn, long oid) throws java.sql.SQLException
     {
         super(conn, oid);
     }
